+++ public/index.php
@@ -4,7 +4,7 @@ use Rarst\ReleaseBelt\Application;
 
 require __DIR__.'/../vendor/autoload.php';
 
-$configPath = __DIR__.'/../config.php';
+$configPath = __DIR__.'/../config/config.php';
 
 $app = new Application(
     file_exists($configPath) ? require $configPath : []